requirements.txt
PyAutoGUI==0.9.52
uvicorn==0.18.3
PyYAML==5.4.1
fastapi==0.85.0
oauthlib==3.1.0
python-socketio==5.7.1
aiohttp==3.8.3
websockets==9.1
requests-oauthlib==1.3.1
jinja2==2.11.3
PyDirectInput==1.0.4
jsonschema==4.16.0